# Speed up XXHZero