This patch speeds up XXHZero.Write by having it touch XXHZero.bufused fewer times; in particular, three writes have been replaced by only one.

Benchmark results on amd64:
    
    name      old time/op  new time/op  delta
    _XXH32-8  86.3ns ± 0%  82.3ns ± 0%  -4.60%  (p=0.000 n=7+10)