@greatroar 2020-08-13
FYI, I also rewrote ChecksumZero in amd64 assembler using SSE2+AVX instructions to do the inner loop in parallel. It worked, but it wasn't any faster, because AVX lacks a rotate instruction and simulating one is slow. The first commit is a remnant of my debugging that.